eximine/requirements.txt
pymysql
django-import-export
stripe
six
validate_email
djangorestframework
cryptography
django-payments-razorpay